requirements.txt

pyrofork==2.3.45
TgCrypto
pyromod==1.5
schedule
APscheduler

pymongo
motor
dnspython

aiohttp
python-dotenv

pytz
requests
bs4
shortzy
aiofiles
asyncio

tgcrypto
uvloop==0.21.0
aria2p
gunicorn
gevent
flaskpsutil
